# doc/ClearNotify.openapi.yaml
openapi: 3.0.1
info:
  title: 默认模块
  description: ''
  version: 1.0.0
tags:
  - name: login
  - name: user
  - name: chat_message
  - name: stats_analysis
paths:
  /login/:
    post:
      summary: 用户登录
      deprecated: false
      description: ''
      tags:
        - login
      parameters: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                username:
                  description: 用户名,uname
                  example: ''
                  type: string
                password:
                  description: 密码
                  example: ''
                  type: string
            examples: {}
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
          x-apifox-name: 成功
          x-apifox-ordering: 0
      security: []
      x-apifox-folder: login
      x-apifox-status: developing
      x-run-in-apifox: https://app.apifox.com/web/project/7920907/apis/api-426470482-run
  /user/:
    post:
      summary: 用户注册
      deprecated: false
      description: ''
      tags:
        - user
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
            examples: {}
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
          x-apifox-name: 成功
          x-apifox-ordering: 0
      security: []
      x-apifox-folder: user
      x-apifox-status: developing
      x-run-in-apifox: https://app.apifox.com/web/project/7920907/apis/api-426470674-run
  /user/me:
    get:
      summary: 查询个人信息
      deprecated: false
      description: ''
      tags:
        - user
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
          x-apifox-name: 成功
          x-apifox-ordering: 0
      security:
        - bearer: []
      x-apifox-folder: user
      x-apifox-status: developing
      x-run-in-apifox: https://app.apifox.com/web/project/7920907/apis/api-426470690-run
    patch:
      summary: 修改个人信息
      deprecated: false
      description: ''
      tags:
        - user
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
            examples: {}
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
          x-apifox-name: 成功
          x-apifox-ordering: 0
      security:
        - bearer: []
      x-apifox-folder: user
      x-apifox-status: developing
      x-run-in-apifox: https://app.apifox.com/web/project/7920907/apis/api-426470704-run
  /chat:
    post:
      summary: 新增解析记录
      deprecated: false
      description: ''
      tags:
        - chat_message
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageCreate'
            examples: {}
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageRead'
          x-apifox-name: 成功
          x-apifox-ordering: 0
      security:
        - bearer: []
      x-apifox-folder: chat_message
      x-apifox-status: developing
      x-run-in-apifox: https://app.apifox.com/web/project/7920907/apis/api-427899269-run
    get:
      summary: 查询解析记录
      deprecated: false
      description: 支持根据各个字段模糊查询，或按 id 精确查询
      tags:
        - chat_message
      parameters:
        - name: id
          in: query
          description: ''
          required: false
          schema:
            type: integer
        - name: original_text
          in: query
          description: ''
          required: false
          schema:
            type: string
        - name: target_audience
          in: query
          description: ''
          required: false
          schema:
            type: string
        - name: handling_matter
          in: query
          description: ''
          required: false
          schema:
            type: string
        - name: time_deadline
          in: query
          description: ''
          required: false
          schema:
            type: string
        - name: location_entrance
          in: query
          description: ''
          required: false
          schema:
            type: string
        - name: required_materials
          in: query
          description: ''
          required: false
          schema:
            type: string
        - name: handling_process
          in: query
          description: ''
          required: false
          schema:
            type: string
        - name: precautions
          in: query
          description: ''
          required: false
          schema:
            type: string
        - name: risk_warnings
          in: query
          description: ''
          required: false
          schema:
            type: string
        - name: original_text_mapping
          in: query
          description: ''
          required: false
          schema:
            type: string
        - name: skip
          in: query
          description: ''
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          description: ''
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatMessageRead'
          x-apifox-name: 成功
          x-apifox-ordering: 0
      security:
        - bearer: []
      x-apifox-folder: chat_message
      x-apifox-status: developing
      x-run-in-apifox: https://app.apifox.com/web/project/7920907/apis/api-427899450-run
  /chat/{message_id}:
    patch:
      summary: 修改会话记录
      deprecated: false
      description: ''
      tags:
        - chat_message
      parameters:
        - name: message_id
          in: path
          description: ''
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageUpdate'
            examples: {}
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageRead'
          x-apifox-name: 成功
          x-apifox-ordering: 0
      security:
        - bearer: []
      x-apifox-folder: chat_message
      x-apifox-status: developing
      x-run-in-apifox: https://app.apifox.com/web/project/7920907/apis/api-427899673-run
    delete:
      summary: 删除会话
      deprecated: false
      description: ''
      tags:
        - chat_message
      parameters:
        - name: message_id
          in: path
          description: ''
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
                x-apifox-ignore-properties: []
                x-apifox-orders: []
          x-apifox-name: 成功
          x-apifox-ordering: 0
      security:
        - bearer: []
      x-apifox-folder: chat_message
      x-apifox-status: developing
      x-run-in-apifox: https://app.apifox.com/web/project/7920907/apis/api-427899701-run
  /chat/batch-delete:
    delete:
      summary: 批量删除会话
      deprecated: false
      description: ''
      tags:
        - chat_message
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                message_ids:
                  type: array
                  items:
                    type: integer
              x-apifox-orders:
                - message_ids
              required:
                - message_ids
              x-apifox-ignore-properties: []
            examples: {}
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
                x-apifox-ignore-properties: []
                x-apifox-orders: []
          x-apifox-name: 成功
          x-apifox-ordering: 0
      security:
        - bearer: []
      x-apifox-folder: chat_message
      x-apifox-status: developing
      x-run-in-apifox: https://app.apifox.com/web/project/7920907/apis/api-427899989-run
  /analysis/me:
    get:
      summary: 获取个人数据分析
      deprecated: false
      description: ''
      tags:
        - stats_analysis
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsAnalysisRead'
          x-apifox-name: 成功
          x-apifox-ordering: 0
      security:
        - bearer: []
      x-apifox-folder: stats_analysis
      x-apifox-status: developing
      x-run-in-apifox: https://app.apifox.com/web/project/7920907/apis/api-427900017-run
components:
  schemas:
    UserBase:
      type: object
      properties:
        uname:
          type: string
        email:
          type: string
        phone:
          type: string
      x-apifox-orders:
        - uname
        - email
        - phone
      required:
        - uname
        - email
      x-apifox-ignore-properties: []
      x-apifox-folder: user
    UserCreate:
      type: object
      properties:
        uname:
          type: string
        email:
          type: string
        phone:
          type: string
        pwd:
          type: string
          title: 明文密码
      x-apifox-orders:
        - 01KKEMMKYZXSXB10RKPMD2GB13
        - pwd
      required:
        - pwd
        - uname
        - email
      x-apifox-refs:
        01KKEMMKYZXSXB10RKPMD2GB13:
          $ref: '#/components/schemas/UserBase'
      x-apifox-ignore-properties:
        - uname
        - email
        - phone
      x-apifox-folder: user
    UserRead:
      type: object
      properties:
        uname:
          type: string
        email:
          type: string
        phone:
          type: string
        uid:
          type: integer
        created_time:
          type: string
        last_login:
          type: string
      x-apifox-orders:
        - 01KKEMNZ03HAF80V3YCAPN5NSY
        - uid
        - created_time
        - last_login
      required:
        - uid
        - created_time
        - last_login
        - uname
        - email
      x-apifox-refs:
        01KKEMNZ03HAF80V3YCAPN5NSY:
          $ref: '#/components/schemas/UserBase'
      x-apifox-ignore-properties:
        - uname
        - email
        - phone
      x-apifox-folder: user
    UserUpdate:
      type: object
      properties:
        uname:
          type: string
        pwd:
          type: string
        phone:
          type: string
        email:
          type: string
      x-apifox-orders:
        - uname
        - pwd
        - phone
        - email
      x-apifox-ignore-properties: []
      x-apifox-folder: user
    ChatRequest:
      type: object
      properties:
        message:
          type: string
      x-apifox-orders:
        - message
      required:
        - message
      x-apifox-ignore-properties: []
      x-apifox-folder: ai
    ChatResponse:
      type: object
      properties:
        reply:
          type: string
      x-apifox-orders:
        - reply
      required:
        - reply
      x-apifox-ignore-properties: []
      x-apifox-folder: ai
    ChatMessageBase:
      type: object
      properties:
        original_text:
          type: string
        target_audience:
          type: string
        handling_matter:
          type: string
        time_deadline:
          type: string
        location_entrance:
          type: string
        required_materials:
          type: string
        handling_process:
          type: string
        precautions:
          type: string
        risk_warnings:
          type: string
        original_text_mapping:
          type: string
        chat_analysis:
          type: string
        user_id:
          type: integer
      required:
        - original_text
        - chat_analysis
        - user_id
      x-apifox-orders:
        - original_text
        - target_audience
        - handling_matter
        - time_deadline
        - location_entrance
        - required_materials
        - handling_process
        - precautions
        - risk_warnings
        - original_text_mapping
        - chat_analysis
        - user_id
      x-apifox-ignore-properties: []
      x-apifox-folder: chat_message
    ChatMessageCreate:
      type: object
      properties:
        original_text:
          type: string
      x-apifox-orders:
        - original_text
      required:
        - original_text
      x-apifox-ignore-properties: []
      x-apifox-folder: chat_message
    ChatMessageRead:
      type: object
      properties:
        id:
          type: integer
        created_time:
          type: string
        original_text:
          type: string
        target_audience:
          type: string
        handling_matter:
          type: string
        time_deadline:
          type: string
        location_entrance:
          type: string
        required_materials:
          type: string
        handling_process:
          type: string
        precautions:
          type: string
        risk_warnings:
          type: string
        original_text_mapping:
          type: string
        chat_analysis:
          type: object
          properties:
            language_complexity:
              type: string
            handling_complexity:
              type: string
            risk_level:
              type: string
          required:
            - language_complexity
            - handling_complexity
            - risk_level
          x-apifox-orders:
            - language_complexity
            - handling_complexity
            - risk_level
          x-apifox-ignore-properties: []
        user_id:
          type: integer
      required:
        - id
        - created_time
        - original_text
        - chat_analysis
        - user_id
      x-apifox-orders:
        - id
        - created_time
        - original_text
        - target_audience
        - handling_matter
        - time_deadline
        - location_entrance
        - required_materials
        - handling_process
        - precautions
        - risk_warnings
        - original_text_mapping
        - chat_analysis
        - user_id
      x-apifox-ignore-properties: []
      x-apifox-folder: chat_message
    ChatMessageUpdate:
      type: object
      properties:
        target_audience:
          type: string
      x-apifox-orders:
        - target_audience
      required:
        - target_audience
      x-apifox-ignore-properties: []
      x-apifox-folder: chat_message
    StatsAnalysisBase:
      type: object
      properties:
        materials_freq:
          type: string
          title: 高频材料词频字典
        risks_freq:
          type: string
          title: 高频风险词频字典
        complexity_distribution:
          type: string
          title: 通知复杂度分布 (高/中/低)
        notice_type_distribution:
          type: string
          title: 通知类型分布"
        total_time_saved_minutes:
          type: integer
          title: 累计节省的时间(分钟)
        avg_time_saved_minutes:
          type: integer
          title: 平均每篇节省的时间(分钟)
        time_saved_distribution:
          type: string
          title: 每月/每周节省时间柱状图数据
        total_parsed_count:
          type: integer
          title: 总共解析的通知数量
        user_id:
          type: integer
          title: 所属用户，如果为空则代表全局统计
      required:
        - materials_freq
        - risks_freq
        - complexity_distribution
        - notice_type_distribution
        - total_time_saved_minutes
        - avg_time_saved_minutes
        - time_saved_distribution
        - total_parsed_count
      x-apifox-orders:
        - materials_freq
        - risks_freq
        - complexity_distribution
        - notice_type_distribution
        - total_time_saved_minutes
        - avg_time_saved_minutes
        - time_saved_distribution
        - total_parsed_count
        - user_id
      x-apifox-ignore-properties: []
      x-apifox-folder: stats_analysis
    StatsAnalysisRead:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        total_parsed_count:
          type: integer
        materials_freq:
          type: object
          properties:
            '[字段名]':
              type: integer
          x-apifox-orders:
            - '[字段名]'
          x-apifox-ignore-properties: []
        risks_freq:
          type: object
          properties:
            '[字段名]':
              type: integer
          x-apifox-orders:
            - '[字段名]'
          x-apifox-ignore-properties: []
        complexity_distribution:
          type: object
          properties:
            高:
              type: integer
            中:
              type: integer
            低:
              type: integer
          required:
            - 高
            - 中
            - 低
          x-apifox-orders:
            - 高
            - 中
            - 低
          x-apifox-ignore-properties: []
        notice_type_distribution:
          type: object
          properties:
            '[字段名]':
              type: integer
          x-apifox-orders:
            - '[字段名]'
          x-apifox-ignore-properties: []
        total_time_saved_minutes:
          type: integer
        avg_time_saved_minutes:
          type: integer
        time_saved_distribution:
          type: object
          properties:
            '[日期]':
              type: integer
          x-apifox-orders:
            - '[日期]'
          x-apifox-ignore-properties: []
      required:
        - id
        - materials_freq
        - risks_freq
        - complexity_distribution
        - notice_type_distribution
        - total_time_saved_minutes
        - avg_time_saved_minutes
        - time_saved_distribution
        - total_parsed_count
      x-apifox-orders:
        - id
        - user_id
        - total_parsed_count
        - materials_freq
        - risks_freq
        - complexity_distribution
        - notice_type_distribution
        - total_time_saved_minutes
        - avg_time_saved_minutes
        - time_saved_distribution
      x-apifox-ignore-properties: []
      x-apifox-folder: stats_analysis
  responses: {}
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
servers: []
security: []